+++ src/test/java/it/com/isroot/stash/plugin/GlobalHookTest.java
@@ -1,5 +1,6 @@
 package it.com.isroot.stash.plugin;
 
+import com.google.common.collect.ImmutableMap;
 import it.com.isroot.stash.plugin.util.YaccRule;
 import org.eclipse.jgit.api.Git;
 import org.eclipse.jgit.transport.PushResult;
@@ -55,6 +56,44 @@ public class GlobalHookTest {
     }
 
     @Test
+    public void testRepoHookUsedInsteadIfEnabled() {
+        gitRepoRule.configureYaccGlobalHook(ImmutableMap
+                .of("commitMessageRegex", "global"));
+
+        gitRepoRule.enableYaccRepoHook();
+        gitRepoRule.configureYaccRepoHook(ImmutableMap
+                .of("commitMessageRegex", "repo"));
+
+        // Push a commit that would have been allowed by global settings. It
+        // should be rejected due to repo hook.
+        PushResult pushResult = gitRepoRule.getGitRepo()
+                .commitFile("file.java", "global")
+                .push();
+        assertThat(pushResult.getRemoteUpdates()).extracting(RemoteRefUpdate::getStatus)
+                .containsExactly(RemoteRefUpdate.Status.REJECTED_OTHER_REASON);
+    }
+
+    @Test
+    public void testGlobalHookUsedAfterRepoHookToggledOnOff() {
+        // Make sure global hook configured
+        gitRepoRule.configureYaccGlobalHook(ImmutableMap
+                .of("commitMessageRegex", "global"));
+
+        // Toggle hook settings on/off
+        gitRepoRule.enableYaccRepoHook();
+        gitRepoRule.configureYaccRepoHook(ImmutableMap
+                .of("commitMessageRegex", "repo"));
+        gitRepoRule.disableYaccRepoHook();
+
+        // Push a commit... this should be rejected by global settings
+        PushResult pushResult = gitRepoRule.getGitRepo()
+                .commitFile("file.java", "repo")
+                .push();
+        assertThat(pushResult.getRemoteUpdates()).extracting(RemoteRefUpdate::getStatus)
+                .containsExactly(RemoteRefUpdate.Status.REJECTED_OTHER_REASON);
+    }
+
+    @Test
     public void testBranchNameCheck_runsBeforeCommitChecks() throws Exception {
         Map<String, String> settings = new HashMap<>();
         settings.put("requireMatchingAuthorEmail", "true");
@@ -81,5 +120,33 @@ public class GlobalHookTest {
         assertThat(pushResult.getMessages())
                 .contains("refs/heads/new-branch: Invalid branch name. 'new-branch' does not match regex 'dev'")
                 .doesNotContain("expected committer");
+    }
+
+    /**
+     * Same idea as {@link #testGlobalHookUsedAfterRepoHookToggledOnOff()}, but spot checks ref
+     * checks. Branch name check is a ref check, not a commit check, so works a little differently.
+     */
+    @Test
+    public void testBranchNameCheck_globalHookUsedAfterRepoHookToggledOnOff() throws Exception {
+        Git git = gitRepoRule.getGitRepo().getGit();
+
+        gitRepoRule.configureYaccGlobalHook(ImmutableMap
+                .of("branchNameRegex", "global"));
+
+        // Toggle hook settings on/off
+        gitRepoRule.enableYaccRepoHook();
+        gitRepoRule.configureYaccRepoHook(ImmutableMap
+                .of("branchNameRegex", "repo"));
+        gitRepoRule.disableYaccRepoHook();
+
+        // Push a branch... this should be rejected by global settings
+        git.branchCreate()
+                .setName("repo")
+                .call();
+
+        PushResult pushResult = gitRepoRule.getGitRepo()
+                .push("repo");
+        assertThat(pushResult.getRemoteUpdates()).extracting(RemoteRefUpdate::getStatus)
+                .containsExactly(RemoteRefUpdate.Status.REJECTED_OTHER_REASON);
     }
 }